requirements.txt
Django == 4.1.2
djangorestframework == 3.14.0
djangorestframework-simplejwt == 5.2.1
Pillow == 9.2.0
psycopg2-binary == 2.9.4
drf-yasg == 1.21.4
google-auth == 2.13.0
facebook-sdk == 3.1.0
python-twitter == 3.5

